Extract and fix query nft info test
