+++ src/lib.rs
@@ -32,8 +32,9 @@ mod unittest_query_contract_config;
 mod unittest_query_contract_info;
 mod unittest_query_minters;
 mod unittest_query_nft_dossier;
-mod unittest_query_owner_of;
+mod unittest_query_nft_info;
 mod unittest_query_num_tokens;
+mod unittest_query_owner_of;
 mod unittest_query_tokens;
 mod unittest_query_unwrapped;
 mod unittest_register_receive_nft;
